docker_compose/movies_db/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
Django==4.1.3
django_cors_headers==3.13.0
django-debug-toolbar==3.7.0
django-split-settings==1.2.0
exceptiongroup==1.0.1
flake8==5.0.4
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
sqlparse==0.4.3
tomli==2.0.1
